services/image-processing/requirements.txt
flask
tensorflow
pillow
numpy